array validator not accepting boolean attributes
when having something like:

    param :recursive, [true, false]

the validator fails (undefined method `TrueClass' for Kernel:Module).

Simplifying the array validator not to try typecast the values for now. Maybe
there could be some option for that in the param to enable something like this
explicitly. Implicit conversions don't fit here very well. If I say I expect a
param to be one of [1, 2, 3] I might not be happy getting string instead of integer.