reload controllers in development environment when showing apidoc
It allows instant reload of documentation content. It reloads all the
controllers only when loading the html part (not JSON), so that it's slowed
down only when refreshing the whole page and not on subsequent JSON calls.

It would be better to make this feature customizable by some setting, but for
now lets keep it stupid and simple: it's turned on only in development
environment.