add exception classes Apipie::ParamInvalid and Apipie::ParamMissing.
These allow controllers to trap validation failures and return, e.g. status 400 with JSON response listing the offending field and value.  (The old way, using ArgumentError, made it hard to distinguish between a server coding error and a client request error.)

For backwards compatibility, these classes are subclasses of ArgumentError, and ArgumentError will be raised if a custom validator returns a string instead of exception.