Remove the client generator from Apipie itself
It's hart to make the client code generation generic enough to fit
all the projects (including authentization, versioning etc.).

Therefore, I'm removing the code itself from this gem. Preferred way
is to use the apipie.json data that this gem provides and use that to
generate the code that fits the project needs.

An example of a client generator using the API is here [1]

[1] - https://github.com/mbacovsky/foreman_api